Enable Sync in Windows builds (#3731)

+++ src/realm/object-store/sync/sync_user.cpp
@@ -65,8 +65,8 @@ RealmJWT::RealmJWT(const std::string& token)
     auto json = static_cast<bson::BsonDocument>(bson::parse(json_str));
 
     this->token = token;
-    this->expires_at = static_cast<int64_t>(json["exp"]);
-    this->issued_at = static_cast<int64_t>(json["iat"]);
+    this->expires_at = long(static_cast<int64_t>(json["exp"]));
+    this->issued_at = long(static_cast<int64_t>(json["iat"]));
 
     if (json.find("user_data") != json.end()) {
         this->user_data = static_cast<bson::BsonDocument>(json["user_data"]);
